bind/requirements.txt
pytest==5.4.1
dnspython==1.16.0
docker==4.2.0
